server/requirements.txt
fastapi
uvicorn
psycopg2-binary
supabase
python-dotenv
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
supabase==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic==2.5.0
pydantic-settings==2.1.0
pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.26.4
scikit-learn==1.3.2
sentence-transformers==2.2.2
openai==1.3.7
aiofiles==23.2.1
httpx==0.24.1
PyMuPDF==1.23.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyPDF2
pydantic[email]